chore: (cdk) Adds initial CDK stack `cdk-init` has been run, with the resulting output modified to be a little personal tastes.
This includes restructuring the file structure, adding eslint,
and configurations for eslint, prettier, and TS.

The project layout that `cdk-init` is fine, though it isn't
structured in a way that makes sense to myself personally.

On top of that, I have some opinions on the use of prettier, eslint, and
jest in CDK projects, so these will be applied here as a starting point.