chore: Adds .envrc_example
A `direnv` `.envrc_example` file has been added to promote ease of
managing environment variables expected within the context of this
project.

Contributers are to create a `.envrc` file, and define their variables
within that. Said file should not be comitted, hence, it has been added
to the `.gitignore` file.