feat: Adds API key to API Gateway
In addition to the above, the resource Ids across the stack have been
correctly named to match the working title.

This change puts down the foundations to support API Key authentication
against the API Gateway. Though for now, the API Key requirement is not
enabled while on client-side work is in progress.