feat: Adds AWS Secrets manager support to Lambda
It was noticed that in the event that a request happened to include the
API Key for the API Gateway, (or any secret of similar nature for that
matter) those secrets could be returned in a curl request if the body
was returned for any reason.

So, as a precaution, a scrubber has been added to remove the API key
provided that the key matches a particular Secrets Manager resource
that is a copy of the API Key value.