chore: Removes request.body from Lambda return
The body doesn't need to be explictly returned in a request when the API
endpoint is given a POST request.

Currently, the `fmt.Printf` still include the request body, though since
these are used for logging in Cloud Watch, these can remain for now.
Albeit, with the API key still being scrubbed in the request body.