chore: Cleans up test config
The test resource `formationId` has been removed, as it was only
there for demonstrative purposes.

The throttle settings fo the individual methods for the plan have been
tweaked to be more generous as we get closer to the game-jame-2022 date.
We will need to open up the amount of usage the API can experience as we
demo the MVP out to friends and family

Additionally, the golang code responsible for this endpoint has been
cleaned up since it no longer serves any purpose.