feat: Adds components for CORS support An `OPTIONS` method has been added to the API, and each response from the API includes the mandatory headers to satisfy CORS requirements.
Without these changes, Cross-Origin-Requests against the API will fail.
The Infrastructure is set up so that the Lambda acts as the proxy for
handling requests. Because of this, the Lambda is responsible for
returning the headers for requests coming in.

Alternatively, API Gateway can reply with CORS headers itself, though it
 would need to be configured to also pass through the Lambda response.
 This implementation proved easier to exercise