refactor: Improves layout of http method code
Each http methods has been broken out into individual functions. This
is so that the functions are easier to maintain rather than working with
a mega-function.

NOTE: When a non-error occurs in the API, the method currently only
returns the generic internal error message and status code. This will be
fixed in the next commit.