fix: golang ci file open errors
This commit fixes an issue where `golangci-lint` throws a series of
errors claiming to unable to open untarred files. This is supposedly
because the tool is supposed to run in isolation in Github actions.
Read:with no other accompanying steps, or cache (what?).

Refer to: https://github.com/golangci/golangci-lint-action/issues/244